pyproject.toml
[tool.poetry]
name = "AI_Hub_LineBot"
version = "0.1.5"
description = "ChatGPT、BingGPT、Image Creator、DALL_E、Other AI Application with Line Bot"
authors = ["Lin-Rexter <aaaa35963596@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ai_together"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.3"
line-bot-sdk = "^2.4.2"
revchatgpt = "^4.0.9"
openai = "^0.27.2"
edgegpt = "^0.1.10.1"
googletrans = "4.0.0rc1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
